fix(ci): fix abstract class error and MSVC warnings in DX12
This fixes remaining CI build failures for Windows:
- Implements IResource::GetType() in DX12Texture, DX12Buffer, DX12Shader, and DX12Sampler to resolve 'cannot instantiate abstract class' errors.
- Fixes warning C4100 'unreferenced parameter' for 'filename' in DX12SwapChain.cpp.
- Fixes warning C4244 'possible loss of data' in DX12Texture.cpp by adding static_cast<UINT16>.

+++ src/engine/graphic/adapters/dx12/DX12Texture.cpp
@@ -442,8 +442,8 @@ D3D12_RESOURCE_DESC DX12Texture::GetD3D12ResourceDesc() const {
     desc.Alignment = 0;
     desc.Width = m_desc.width;
     desc.Height = m_desc.height;
-    desc.DepthOrArraySize = m_desc.arraySize;
-    desc.MipLevels = m_desc.mipLevels;
+    desc.DepthOrArraySize = static_cast<UINT16>(m_desc.arraySize);
+    desc.MipLevels = static_cast<UINT16>(m_desc.mipLevels);
     desc.Format = GetDXGIFormat();
     desc.SampleDesc.Count = 1;
     desc.SampleDesc.Quality = 0;
